requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
exceptiongroup==1.3.1
fastapi==0.132.0
greenlet==3.3.2
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
